mysite/requirements.txt
Django==3.2.13
djangorestframework==3.13.1
django-filter==21.1
coverage==6.4.1
requests
xmltodict